Edited Admin.ManageUsers.ManageEmployee.editemployee and Admin.ManageUsers.ManageEmployee.editemployee.ItemTemplate1; Deleted Admin.ManageUsers.ManageEmployee.editemployee.Form1, Admin.ManageUsers.ManageEmployee.editemployee.Form1.ItemTemplate2, Admin.ManageUsers.ManageEmployee.editemployee.ItemTemplate4 and Admin.ManageUsers.ManageEmployee.editemployee.ItemTemplate3
