--- client_code/Admin/ManageUsers/ManageEmployee/editemployee/Form1/ItemTemplate2/__init__.py
@@ -1,10 +1,0 @@
-from ._anvil_designer import ItemTemplate2Template
-from anvil import *
-
-
-class ItemTemplate2(ItemTemplate2Template):
-  def __init__(self, **properties):
-    # Set Form properties and Data Bindings.
-    self.init_components(**properties)
-
-    # Any code you write here will run before the form opens.
