--- client_code/Admin/ManageUsers/ManageEmployee/editemployee/Form1/ItemTemplate2/form_template.yaml
@@ -1,3 +1,0 @@
-is_package: true
-container: {type: ColumnPanel}
-components: []
