--- client_code/Admin/ManageUsers/ManageEmployee/editemployee/Form1/__init__.py
@@ -1,21 +1,0 @@
-from ._anvil_designer import Form1Template
-from anvil import *
-
-
-class Form1(Form1Template):
-  def __init__(self, **properties):
-    # Set Form properties and Data Bindings.
-    self.init_components(**properties)
-    # self.repeating_panel_1.items=app_tables.employee.search(
-
-    # Any code you write here will run before the form opens.
-
-  def button_1_click(self, **event_args):
-    """This method is called when the button is clicked"""
-    # app_tables.employee.add_row()
-
-  def link_1_click(self, **event_args):
-    """This method is called when the link is clicked"""
-    open_form('Admin.ManageUsers.ManageAdmin.editadmin')
-
-
