--- client_code/Admin/ManageUsers/ManageEmployee/editemployee/Form1/form_template.yaml
@@ -1,14 +1,0 @@
-is_package: true
-container:
-  type: HtmlTemplate
-  properties: {html: '@theme:standard-page.html'}
-components:
-- type: FlowPanel
-  properties: {}
-  name: navbar_links
-  layout_properties: {slot: nav-right}
-  components: []
-- type: Label
-  properties: {role: null, bold: false, align: left, spacing_above: small, background: '', icon: '', underline: false, font: '', icon_align: left, foreground: '', spacing_below: small, visible: true, border: '', italic: false, font_size: null, tooltip: '', text: EDIT EMPLOYEE}
-  name: label_1
-  layout_properties: {slot: title}
