+++ client_code/Admin/ManageUsers/ManageEmployee/editemployee/ItemTemplate1/form_template.yaml
@@ -1,3 +1,34 @@
-container: {type: ColumnPanel}
-components: []
+container:
+  type: ColumnPanel
+  properties: {col_widths: '{}'}
+components:
+- type: TextBox
+  properties: {}
+  name: text_box_1
+  layout_properties: {grid_position: 'YIOBMY,PBELRQ'}
+  event_bindings: {pressed_enter: text_box_1_pressed_enter}
+  data_bindings:
+  - {property: text, code: 'self.item[''full_name'']', writeback: true}
+- type: TextBox
+  properties: {}
+  name: text_box_2
+  layout_properties: {grid_position: 'YIOBMY,LMREOF'}
+  data_bindings:
+  - {property: text, code: 'self.item[''email_user'']', writeback: true}
+- type: TextBox
+  properties: {}
+  name: text_box_3
+  layout_properties: {grid_position: 'YIOBMY,WKQTQK'}
+  data_bindings:
+  - {property: text, code: 'self.item[''user_phonenumber'']', writeback: true}
+- type: Button
+  properties: {role: primary-color, bold: false, align: center, spacing_above: small, background: '', icon: '', underline: false, font: '', icon_align: left, foreground: '', spacing_below: small, visible: true, border: '', italic: false, enabled: true, font_size: null, tooltip: '', text: save}
+  name: button_1
+  layout_properties: {grid_position: 'YIOBMY,NNZWAN'}
+  event_bindings: {click: button_1_click}
+- type: Button
+  properties: {role: primary-color, bold: false, align: center, spacing_above: small, background: '', icon: '', underline: false, font: '', icon_align: left, foreground: '', spacing_below: small, visible: true, border: '', italic: false, enabled: true, font_size: null, tooltip: '', text: delete}
+  name: button_2
+  layout_properties: {grid_position: 'YIOBMY,QWIPBJ'}
+  event_bindings: {click: button_2_click}
 is_package: true
