--- client_code/Admin/ManageUsers/ManageEmployee/editemployee/ItemTemplate3/__init__.py
@@ -1,10 +1,0 @@
-from ._anvil_designer import ItemTemplate3Template
-from anvil import *
-
-
-class ItemTemplate3(ItemTemplate3Template):
-  def __init__(self, **properties):
-    # Set Form properties and Data Bindings.
-    self.init_components(**properties)
-
-    # Any code you write here will run before the form opens.
