--- client_code/Admin/ManageUsers/ManageEmployee/editemployee/ItemTemplate3/form_template.yaml
@@ -1,14 +1,0 @@
-container:
-  type: ColumnPanel
-  properties: {col_widths: '{}'}
-components:
-- type: Link
-  properties: {col_widths: '{}'}
-  name: link_1
-  layout_properties: {grid_position: 'KHVXBK,RBVBEE'}
-  components:
-  - type: Button
-    properties: {text: button_1}
-    name: button_1
-    layout_properties: {grid_position: 'UTIXCH,WSZTST'}
-is_package: true
