--- client_code/Admin/ManageUsers/ManageEmployee/editemployee/ItemTemplate4/__init__.py
@@ -1,24 +1,0 @@
-from ._anvil_designer import ItemTemplate4Template
-from anvil import *
-
-
-class ItemTemplate4(ItemTemplate4Template):
-  def __init__(self, **properties):
-    # Set Form properties and Data Bindings.
-    self.init_components(**properties)
-
-    # Any code you write here will run before the form opens.
-
-  def link_1_click(self, **event_args):
-    """This method is called when the link is clicked"""
-    pass
-
-  def button_2_click(self, **event_args):
-    """This method is called when the button is clicked"""
-    self.item.delete()
-    self.remove_from_parent()
-
-  def button_1_click(self, **event_args):
-    """This method is called when the button is clicked"""
-    open_form('Admin.ManageUsers.ManageEmployee.editemployee.Form1')
-
