--- client_code/Admin/ManageUsers/ManageEmployee/editemployee/ItemTemplate4/form_template.yaml
@@ -1,22 +1,0 @@
-container:
-  type: ColumnPanel
-  properties: {col_widths: '{}'}
-components:
-- type: Link
-  properties: {}
-  name: link_1
-  layout_properties: {grid_position: 'KVOQGQ,HLAQWP'}
-  event_bindings: {click: link_1_click}
-  data_bindings:
-  - {property: text, code: 'self.item[''full_name'']'}
-- type: Button
-  properties: {role: null, bold: false, align: center, spacing_above: small, background: '', icon: '', underline: false, font: '', icon_align: left, foreground: '', spacing_below: small, visible: true, border: '', italic: false, enabled: true, font_size: null, tooltip: '', text: edit}
-  name: button_1
-  layout_properties: {grid_position: 'KVOQGQ,QUQICW'}
-  event_bindings: {click: button_1_click}
-- type: Button
-  properties: {role: null, bold: false, align: center, spacing_above: small, background: '', icon: '', underline: false, font: '', icon_align: left, foreground: '', spacing_below: small, visible: true, border: '', italic: false, enabled: true, font_size: null, tooltip: '', text: delete}
-  name: button_2
-  layout_properties: {grid_position: 'KVOQGQ,OJTCVF'}
-  event_bindings: {click: button_2_click}
-is_package: true
