+++ client_code/Admin/ManageUsers/ManageEmployee/editemployee/form_template.yaml
@@ -7,8 +7,17 @@ components:
   properties: {}
   name: navbar_links
   layout_properties: {slot: nav-right}
-  components: []
+  components:
+  - type: Link
+    properties: {role: null, bold: false, align: left, col_widths: '', spacing_above: small, background: '', icon: 'fa:arrow-circle-left', wrap_on: mobile, underline: false, font: '', col_spacing: medium, icon_align: left, foreground: '', url: '', spacing_below: small, visible: true, border: '', italic: false, font_size: null, tooltip: '', text: Manage employe}
+    name: link_1
+    layout_properties: {}
+    event_bindings: {click: link_1_click}
 - type: Label
   properties: {role: null, bold: false, align: center, spacing_above: small, background: '', icon: '', underline: false, font: '', icon_align: left, foreground: '', spacing_below: small, visible: true, border: '', italic: false, font_size: 25, tooltip: '', text: employee}
   name: label_1
   layout_properties: {slot: title}
+- type: RepeatingPanel
+  properties: {item_template: Admin.ManageUsers.ManageEmployee.editemployee.ItemTemplate1}
+  name: repeating_panel_1
+  layout_properties: {slot: default}
